requirements.txt
Django==4.2
djangorestframework==3.14
redis==5.0.0
django-redis==5.3.0
psycopg2==2.9.6
pyTelegramBotAPI==4.12.0